+++ changelog.md
@@ -16,7 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Other work
 -->
 
-## vNEXT - unreleased
+## v2.0.1 - 2026-07-30
 
 ### Fixed